Fix situation when AI tries to move pawn on user behalf in turn; repackage Position; expose method to get pawn to beat
